requirements.txt
pydantic
httpx
aiolimiter
python-dotenv